requirements.txt
Flask
torch
torchvision
pillow
numpy
opencv-python
